[#90] Improve project structure
