+++ src/Universum/Base.hs
@@ -6,14 +6,37 @@
 -- package.
 
 module Universum.Base
-       ( module GHC.Base
+       ( -- * Base types
+         module Data.Bits
+       , module Data.Char
+       , module Data.Int
+       , module Data.Word
+
+         -- * Base type classes
+       , module Data.Eq
+       , module Data.Foldable
+       , module Data.Ord
+       , module Data.Traversable
+
+         -- * System IO
+       , module System.IO
+
+         -- * Base GHC types
+#if ( __GLASGOW_HASKELL__ >= 710 )
+       , module Data.Proxy
+       , module Data.Typeable
+       , module Data.Void
+#endif
+       , module GHC.Base
        , module GHC.Enum
        , module GHC.Err
        , module GHC.Exts
        , module GHC.Float
+       , module GHC.Generics
        , module GHC.Num
        , module GHC.Real
        , module GHC.Show
+
 #if MIN_VERSION_base(4,10,0)
        , module GHC.TypeNats
 #else
@@ -30,12 +53,34 @@ module Universum.Base
        , ($!)
        ) where
 
+-- Base types
+import Data.Bits (xor)
+import Data.Char (chr)
+import Data.Int (Int, Int16, Int32, Int64, Int8)
+import Data.Word (Word, Word16, Word32, Word64, Word8, byteSwap16, byteSwap32, byteSwap64)
+
+-- IO
+import System.IO (FilePath, Handle, IOMode (..), stderr, stdin, stdout, withFile)
+
+-- Base typeclasses
+import Data.Eq (Eq (..))
+import Data.Foldable (Foldable, concat, concatMap, foldlM, foldrM, maximumBy, minimumBy)
+import Data.Ord (Down (..), Ord (..), Ordering (..), comparing)
+import Data.Traversable (Traversable (..), fmapDefault, foldMapDefault, forM, mapAccumL, mapAccumR)
+
 -- Base GHC types
+#if ( __GLASGOW_HASKELL__ >= 710 )
+import Data.Proxy (Proxy (..))
+import Data.Typeable (Typeable)
+import Data.Void (Void, absurd, vacuous)
+#endif
+
 import GHC.Base (String, asTypeOf, maxInt, minInt, ord, seq, (++))
 import GHC.Enum (Bounded (..), Enum (..), boundedEnumFrom, boundedEnumFromThen)
 import GHC.Err (error, undefined)
 import GHC.Exts (Constraint, FunPtr, Ptr)
 import GHC.Float (Double (..), Float (..), Floating (..), showFloat, showSignedFloat)
+import GHC.Generics (Generic)
 import GHC.Num (Integer, Num (..), subtract)
 import GHC.Real hiding ((%))
 import GHC.Show (Show (..))
