+++ src/Universum/Functor.hs
@@ -1,23 +1,11 @@
-{-# LANGUAGE CPP  #-}
 {-# LANGUAGE Safe #-}
 
 -- | Convenient functions to work with 'Functor'.
 
 module Universum.Functor
-       ( Functor (..)
-       , void
-       , ($>)
-       , (<$>)
-       , (<<$>>)
+       ( module Universum.Functor.Compose
+       , module Universum.Functor.Reexport
        ) where
 
-import Data.Function ((.))
-import Data.Functor (Functor (..), void, ($>), (<$>))
-
--- | Alias for @fmap . fmap@. Convenient to work with two nested 'Functor's.
---
--- >>> negate <<$>> Just [1,2,3]
--- Just [-1,-2,-3]
-(<<$>>) :: (Functor f, Functor g) => (a -> b) -> f (g a) -> f (g b)
-(<<$>>) = fmap . fmap
-infixl 4 <<$>>
+import Universum.Functor.Compose
+import Universum.Functor.Reexport
