+++ src/Universum/Functor/Reexport.hs
@@ -1,0 +1,13 @@
+{-# LANGUAGE Safe #-}
+
+module Universum.Functor.Reexport
+       ( module Control.Arrow
+       , module Data.Bifunctor
+       , module Data.Functor
+       , module Data.Functor.Identity
+       ) where
+
+import Control.Arrow ((&&&))
+import Data.Bifunctor (Bifunctor (..))
+import Data.Functor (Functor (..), void, ($>), (<$>))
+import Data.Functor.Identity (Identity (..))
