+++ universum.cabal
@@ -31,31 +31,42 @@ library
                        Universum.Applicative
                        Universum.Base
                        Universum.Bool
+                           Universum.Bool.Guard
+                           Universum.Bool.Reexport
                        Universum.Container
                            Universum.Container.Class
                            Universum.Container.Reexport
                        Universum.Debug
-                       Universum.Exceptions
+                       Universum.DeepSeq
+                       Universum.Exception
+                       Universum.Function
                        Universum.Functor
-                       Universum.List
-                       Universum.Nub
-                       Universum.Print
-                       Universum.String
-                       Universum.TypeOps
-                       Universum.Unsafe
-                       Universum.VarArg
-
+                           Universum.Functor.Compose
+                           Universum.Functor.Reexport
                        Universum.Lifted
                            Universum.Lifted.Concurrent
                            Universum.Lifted.Env
                            Universum.Lifted.File
                            Universum.Lifted.IORef
                            Universum.Lifted.ST
-
+                       Universum.List
+                           Universum.List.Reexport
+                           Universum.List.Safe
                        Universum.Monad
-                         Universum.Monad.Either
-                         Universum.Monad.Maybe
-                         Universum.Monad.Trans
+                           Universum.Monad.Container
+                           Universum.Monad.Either
+                           Universum.Monad.Maybe
+                           Universum.Monad.Reexport
+                           Universum.Monad.Trans
+                       Universum.Monoid
+                       Universum.Nub
+                       Universum.Print
+                       Universum.String
+                           Universum.String.Conversion
+                           Universum.String.Reexport
+                       Universum.TypeOps
+                       Universum.Unsafe
+                       Universum.VarArg
 
   ghc-options:         -Wall -fwarn-implicit-prelude
